requirements.txt
tensorflow-gpu==2.0.0-beta1
tqdm
scipy
pandas
matplotlib
librosa
